Re-implement Ctrl+Shift+Left/Right to move the current tab left or right.  Fix float->int conversion warning.
svn path=/trunk/KDE/kdebase/apps/konsole/; revision=682750

+++ src/ViewSplitter.cpp
@@ -58,7 +58,7 @@ void ViewSplitter::adjustContainerSize(ViewContainer* container , int percentage
     QList<int> containerSizes = sizes();
 
     int oldSize = containerSizes[containerIndex];
-    int newSize = oldSize * ( 1.0 + percentage/100.0 );
+    int newSize = (int)(oldSize * ( 1.0 + percentage/100.0 ));
 
    // qDebug() << "Old container size:" << oldSize << ", new size:" << newSize;
